chapter 11

+++ Exercises/Chapter 10/Radians.java
@@ -2,8 +2,9 @@
 
 public class Radians{
    public static void main (String [] args){
-      int degrees = 30;
+      byte degrees = 30;
       double rad = degrees * Math.PI/180;
       System.out.println("Degrees: " + degrees + " Radians: " + rad);
    }
 }
+
